requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
pymupdf
wikipedia
langchain-text-splitters
requests
arxiv
langchain-openai
sentence-transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_chroma
duckdb
pandas
langchain_groq
langchain_core
sse_starlette
langchain_community
duckduckgo_search
pymysql
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
gradio
